parseInt changed to parseFloat so the user can add decimal numbers

+++ assets/scripts/app.js
@@ -20,8 +20,8 @@ let operator = "";
 let displayValue = "";
 
 function operate(operator, num1, num2) {
-  num1 = parseInt(num1);
-  num2 = parseInt(num2);
+  num1 = parseFloat(num1);
+  num2 = parseFloat(num2);
   let result = 0;
 
   if (operator === "+") {
